+++ library/cc/engine_builder.cc
@@ -124,8 +124,9 @@ EngineBuilder& EngineBuilder::enableSocketTagging(bool socket_tagging_on) {
   return *this;
 }
 
-EngineBuilder& EngineBuilder::usePlatformCertValidator(bool use_platform_cert_validator) {
-  this->use_platform_cert_validator_ = use_platform_cert_validator;
+EngineBuilder&
+EngineBuilder::enablePlatformCertificatesValidation(bool platform_certificates_validation_on) {
+  this->platform_certificates_validation_on_ = platform_certificates_validation_on;
   return *this;
 }
 
@@ -209,8 +210,8 @@ std::string EngineBuilder::generateConfigStr() {
   // Choose the right certification validator.
   absl::StrReplaceAll(
       {{"{{custom_cert_validation_context}}",
-        (this->use_platform_cert_validator_ ? platform_cert_validation_context_template
-                                            : default_cert_validation_context_template)}},
+        (this->platform_certificates_validation_on_ ? platform_cert_validation_context_template
+                                                    : default_cert_validation_context_template)}},
       &config_str);
 
   if (config_str.find("{{") != std::string::npos) {