+++ library/java/org/chromium/net/impl/NativeCronetEngineBuilderImpl.java
@@ -62,7 +62,7 @@ public class NativeCronetEngineBuilderImpl extends CronetEngineBuilderImpl {
   private String mAppId = "unspecified";
   private TrustChainVerification mTrustChainVerification = VERIFY_TRUST_CHAIN;
   private String mVirtualClusters = "[]";
-  private boolean mUsePlatformCertValidator = true;
+  private boolean mEnablePlatformCertificatesValidation = true;
 
   /**
    * Builder for Native Cronet Engine. Default config enables SPDY, disables QUIC and HTTP cache.
@@ -116,6 +116,6 @@ public class NativeCronetEngineBuilderImpl extends CronetEngineBuilderImpl {
         mH2ExtendKeepaliveTimeout, mH2RawDomains, mMaxConnectionsPerHost, mStatsFlushSeconds,
         mStreamIdleTimeoutSeconds, mPerTryIdleTimeoutSeconds, mAppVersion, mAppId,
         mTrustChainVerification, mVirtualClusters, nativeFilterChain, platformFilterChain,
-        stringAccessors, keyValueStores, mUsePlatformCertValidator);
+        stringAccessors, keyValueStores, mEnablePlatformCertificatesValidation);
   }
 }