+++ docs/root/intro/version_history.rst
@@ -15,6 +15,7 @@ Bugfixes:
 
 Features:
 
+- api: add option to support platform provided certificates validation interfaces. (:issue `#2144 <2144>`)
 - kotlin: add a way to tell Envoy Mobile to respect system proxy settings by calling an ``enableProxying(true)`` method on the engine builder. (:issue:`#2416 <2416>`)
 
 
@@ -61,7 +62,6 @@ Features:
 - android: create simple persistent SharedPreferencesStore (:issue: `#2319 <2319>`)
 - iOS: A documentation archive is now included in the GitHub release artifact (:issue: `#2335 <2335>`)
 - api: improved C++ APIs compatibility with Java / Kotlin / Swift (:issue `#2362 <2362>`)
-- api: add option to support platform provided certificates validation interfaces. (:issue `#2144 <2144>`)
 - Android: default to use a ``getaddrinfo``-based system DNS resolver instead of c-ares (:issue: `#2419 <2419>`)
 - iOS: add ``KeyValueStore`` protocol conformance to ``UserDefaults`` (:issue: `#2452 <2452>`)
 - iOS: add experimental option to force all connections to use IPv6. (:issue: `#2396 <2396>`)